cqerl_client: Fix sleep-state message fragment handling
The the sleep-state clause of cqerl_client:handle_info/3 did not account
for the possiblity that BinaryMsg can contain the final fragment of one
message and the initial fragment of the next.  In this situation, it will
discard the initial fragment, causing the next fragment to become stuck
at the front of #client_state.delayed, and preventing the client from
decoding any more frames.