Handle the compression type consistently
The compression type was declared as a `binary()` in some places and as an `atom()` in others. In practice, it was almost always an atom. Create a new type for it and make sure it is handled consistently as an atom.

This necessitated changing `encode_proplist_to_map` so that it can handle atoms as the value type in a proplist.