Fix the return type of close_client
`no_return()` is a signal to dialyzer that the function does not return; either it raises an exception or it is meant to run forever (like the receive loop in a gen_server). This means that any function that calls close_client also never returns, and any code after the call to close_client doesn't execute. This is obviously not what we want.