Change prepared query cache to be per-node rather than per-client
It turns out that C* prepared queries are cached per node and not
discarded when the requesting connection is dropped (but rather only
when the node is restarted or they are cycled out of the cache). That
means we can shrink the size of our prepared query cache so that it only
holds at most one copy of each query per node (with the side benefit of
also only needing to prepare each query once per node).

This also has the benefit of working around the C* bug described at
https://issues.apache.org/jira/browse/CASSANDRA-10786, since we only
cache the query once and remove that instance from the cache if we get
'unprepared'.