Add timeout to add_nodes/1,2,3 gen_server calls
add_nodes/1,2,3 was previously async, but were changed to use gen_server:call
instead to prevent a race condition (#111). That change did however introduce
the possibility of gen_server timing out before the add_to_cluster handler had
completed.

This commit introduces an application env variable that makes it possible to
specify the timeout, and sets it to 30 seconds if not specified.